New Structure for Database Cache(.icondb) Files
This branch will try to maintain database cache in such a way, that forecasts
over a given timespan in the future will be stored or overwritten in same
filenames. E.g. for model run at 00 and forecasted hour +12, the forecast
will be stored at forecsast-YYYYMMDD12.icondb. A later model run at 06 which
forecasted hour +06 will be stored in the same file, which overwrites the
previous one.

This new strategy will be benefical to servers, which would run continously
while updating runs in the background.